Move stuff to room.py
